Find all subgraphs in pairs for every subset of edges (c++ version)

Given a dataframe of edges with strength between nodes this function returns info on every subgraph state achieved by adding nodes in one-at-a-time in descending order of strength.

explore_subgraph_structure(association_pairs, strength_column = "strength")

Arguments

association_pairs
dataframe with columns a and b representing the ids of the variables or nodes and columns strength that is a numeric indicator of strength of association (higher = stronger).

Value

Dataframe with the following columns for the subgraph info

Examples

virus_net %>%
  dplyr::arrange(dplyr::desc(strength)) %>%
    head(1000) %>%
    explore_subgraph_structure()

#> # A tibble: 327 x 11
#>     step n_edges strength n_nodes_seen n_subgraphs max_size rel_max_size
#>    <int>   <int>    <dbl>        <int>       <int>    <int>        <dbl>
#>  1     1       1   0.0916            2           1        2        1    
#>  2     2       2   0.0836            3           1        3        1    
#>  3     3       3   0.0750            5           2        3        0.6  
#>  4     4       4   0.0729            6           2        3        0.5  
#>  5     5       5   0.0728            6           2        3        0.5  
#>  6     6       6   0.0704            7           2        4        0.571
#>  7     7       7   0.0689            8           2        4        0.5  
#>  8     8       8   0.0676            9           2        5        0.556
#>  9     9      10   0.0651           12           3        5        0.417
#> 10    10      11   0.0623           14           4        5        0.357
#> # … with 317 more rows, and 4 more variables: avg_size <dbl>,
#> #   avg_density <dbl>, n_triples <int>, subgraphs <list>