Interactive association subgraphs visualization

Produces an interactive plot of network structure for all possible subgraph arrangements for passed association pairs with summary statistics to guide exploration of cut-points.

visualize_subgraph_structure(
  association_pairs,
  node_info,
  subgraph_results,
  trim_subgraph_results = TRUE,
  warn_of_mismatches = TRUE,
  width = "100%",
  height = "800px"
)

Arguments

association_pairs
dataframe with columns a and b representing the ids of the variables or nodes and columns strength that is a numeric indicator of strength of association (higher = stronger).

node_info
Optional dataframe that has a column id that corresponds to the variables codded in a and b of association_pairs that contains additional info nade available on hover in visualization.

subgraph_results
Dataframe of subgraph results as returned by calculate_subgraph_structure. If it isnt provided it is calculated. Automatic calculation will slow down code depending on how large dataset is.

trim_subgraph_results
Should subgraph results after a giant subgraph has taken over be filtered out? Rule for filtering is at least 10% of the variables are in subgraphs and largest subgraph contains less than 95% of all variables in subgraphs. Allows for easier investigating of the subgraph structure over strength.

warn_of_mismatches
If there are differences in the ids present in association_pairs and node_info should a warning be issued?

width, height
Valid css units for output size (e.g. pixels (px) or percent(%)).

Value

Interactive javascript visualization of association network subgraphs at all possible cut-points

Examples

node_info <- dplyr::rename(virus_host_viruses, id = virus_id)
node_info$color <- ifelse(node_info$type == "RNA", "orangered", "steelblue")
visualize_subgraph_structure(
  virus_net,
  node_info = node_info
)

#> Calculating subgraph structure results...

#> ...finished