Simulate a sticky association network

Used for testing properties of algorithms. Will create an association network that is constructed where some nodes are more "sticky" than others. Or they are in general more associated than others.

simulate_sticky_association_network(
  n_variables = 50,
  n_clusters = 4,
  cluster_coherence = 2,
  stickiness_dist = function(n) {
     stats::rbeta(n, shape1 = 1, shape2 = 4)
 },
  association_dist = function(n) {
     stats::rbeta(n, shape1 = 0.5, shape2 = 10)
 }
)

Arguments

n_variables

How many variables will be simulated

n_clusters

Controls how many clusters are randomly assigned to nodes

cluster_coherence

Scalar that multiplies random association value for nodes in the same clusters. E.g. A value of 2 means variables in same cluster have on average twice the association strength. Most likely above 1.

stickiness_dist

Function that takes a single input: n that will then generate n values from a distribution used to assign stickiness level for each variable. These values will be scaled to an average of 1 so stickiness doesn't change association distribution too much.

association_dist

Function of same format as stickiness_dist that instead is used to generate raw associations between pairs of nodes. The values will then be scaled by cluster_coherence and each variable's stickiness.

Value

List with variables with info about each variable (id, stickiness scalar, and cluster id) and associations which contains the simulated associations in the format that works for all the functions that take association_pairs as an input.

Examples

# Simulate an association network with flat stickiness profile
simulate_sticky_association_network(
  n_variables = 10,
  n_clusters = 3,
  cluster_coherence = 2,
  stickiness_dist = function(n) {runif(n)},
  association_dist = function(n) {rbeta(n, shape1 = 1, shape2 = 10)}
)
#> $variables
#> # A tibble: 10 × 3
#>       id stickiness cluster
#>    <int>      <dbl>   <int>
#>  1     1    1.37          2
#>  2     2    0.593         1
#>  3     3    0.889         2
#>  4     4    1.73          3
#>  5     5    0.647         1
#>  6     6    1.64          3
#>  7     7    1.60          1
#>  8     8    1.18          3
#>  9     9    0.00171       3
#> 10    10    0.355         2
#> 
#> $associations
#> # A tibble: 45 × 3
#>        b same_cluster strength
#>    <int> <lgl>           <dbl>
#>  1     2 FALSE          0.368 
#>  2     3 TRUE           0.172 
#>  3     4 FALSE          0.637 
#>  4     5 FALSE          0.0722
#>  5     6 FALSE          0.333 
#>  6     7 FALSE          0.688 
#>  7     8 FALSE          0.187 
#>  8     9 FALSE          0.211 
#>  9    10 TRUE           0.625 
#> 10     3 FALSE          0.293 
#> # … with 35 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#>